Re-organised display text
I've reorganised the display text into 5 independent files as the single file was getting too complicated to make progress
