+++ Assets/Scripts/ScriptObjects/Afternoon_Text.cs
@@ -1,0 +1,7 @@
+using UnityEngine;
+
+[CreateAssetMenu(fileName = "Afternoon_Text", menuName = "Scriptable Objects/Afternoon_Text")]
+public class Afternoon_Text : ScriptableObject
+{
+    
+}
