+++ Assets/Scripts/ScriptObjects/Afternoon_Text.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: ea6ee9f587d2f6f4688d8606ffedf005
