+++ Assets/Scripts/ScriptObjects/All_Driving_Text.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 604c3fe82988a4544ad8efdb960121c0
