+++ Assets/Scripts/ScriptObjects/All_Work_Text.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b61461742385b2540a1268bca7816951
