--- Assets/Scripts/ScriptObjects/Display_Text.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: e4747e7b7b514a946a570671bf66d62f
