+++ Assets/Scripts/ScriptObjects/Lunch_Text.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: e0cdb687feeb93342bb3dacdfe9275f2
