+++ Assets/Scripts/ScriptObjects/Other_Text.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 17240c81622164541a6c3f8b35edec73
