+++ Assets/Scripts/ScriptObjects/SpawnManagers/Afternoon_Text_Manager.asset
@@ -9,7 +9,6 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e4747e7b7b514a946a570671bf66d62f, type: 3}
-  m_Name: Display_Text_Manager
-  m_EditorClassIdentifier: Assembly-CSharp::Display_Text
-  testText: TESTING text
+  m_Script: {fileID: 11500000, guid: ea6ee9f587d2f6f4688d8606ffedf005, type: 3}
+  m_Name: Afternoon_Text_Manager
+  m_EditorClassIdentifier: Assembly-CSharp::Afternoon_Text
