+++ Assets/Scripts/ScriptObjects/SpawnManagers/Afternoon_Text_Manager.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2450a858a9aeb0140b0eb618064f121a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
