+++ Assets/Scripts/ScriptObjects/SpawnManagers/All_Driving_Text_Manager.asset
@@ -1,0 +1,14 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 604c3fe82988a4544ad8efdb960121c0, type: 3}
+  m_Name: All_Driving_Text_Manager
+  m_EditorClassIdentifier: Assembly-CSharp::All_Driving_Text
