+++ Assets/Scripts/ScriptObjects/SpawnManagers/All_Driving_Text_Manager.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d1da7786d694e94479bed6401eb695d5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
