+++ Assets/Scripts/ScriptObjects/SpawnManagers/All_Work_Text_Manager.asset
@@ -1,0 +1,14 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: b61461742385b2540a1268bca7816951, type: 3}
+  m_Name: All_Work_Text_Manager
+  m_EditorClassIdentifier: Assembly-CSharp::All_Work_Text
