+++ Assets/Scripts/ScriptObjects/SpawnManagers/All_Work_Text_Manager.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 39d815d205ace5e47bc221e053679c0a
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
