+++ Assets/Scripts/ScriptObjects/SpawnManagers/Lunch_Text_Manager.asset
@@ -1,0 +1,14 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e0cdb687feeb93342bb3dacdfe9275f2, type: 3}
+  m_Name: Lunch_Text_Manager
+  m_EditorClassIdentifier: Assembly-CSharp::Lunch_Text
