+++ Assets/Scripts/ScriptObjects/SpawnManagers/Lunch_Text_Manager.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: b233c956c3e53e74c9031b639c2efef7
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
