+++ Assets/Scripts/ScriptObjects/SpawnManagers/Other_Text_Manager.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d93629fe28d315142ba53a427d69c0a8
+guid: fcf75994d0c5dc64eaf9050bd499ae32
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
